requirements.txt
streamlit>=1.30.0
langchain>=0.2.0,<0.3.0
langchain-anthropic>=0.1.0,<0.2.0
langchain-core>=0.2.0,<0.3.0
python-dotenv>=1.0.0
